+++ HISTORY.rst
@@ -4,6 +4,7 @@
 - Add 'copy' functionality to update_object endpoint. In this case the view accepts the (augeias) location
 of the file object and copies it to a new location. (#17)
 - Add Endpoint to create object and key. (#20)
+- Add Endpoint to retrieve meta data (mimetype, size, time last modification) from object.
 
 0.2.0 (07-11-2015)
 ------------------